+++ testing/unit_tests/test_model.py
@@ -1,8 +1,8 @@
-import ivadomed.architecture.blocks
+import ivadomed.architecture.block.film_generator
 import ivadomed.architecture.densenet
 import ivadomed.architecture.resnet
 import ivadomed.architecture.unet
-import ivadomed.architecture.unet3d
+import ivadomed.architecture.unet.unet3d
 import ivadomed.models as imed_model
 import torch
 import torchvision
@@ -20,7 +20,7 @@ def test_model_3d_att():
     # verifying if 3d attention model can be created
     a = [[[[[0 for i in range(48)] for j in range(48)] for k in range(16)]]]
     inp = torch.tensor(a).float()
-    model = ivadomed.architecture.unet3d.Modified3DUNet(in_channel=1, out_channel=1, attention=True)
+    model = ivadomed.architecture.unet.unet3d.Modified3DUNet(in_channel=1, out_channel=1, attention=True)
     inf = model(inp)
     assert(type(inf) == torch.Tensor)
 
@@ -44,7 +44,7 @@ def test_densenet():
 def test_filmed_unet():
     a = [[[[0 for i in range(100)] for i in range(100)]]]
     inp = torch.tensor(a).float()
-    model = ivadomed.architecture.unet.FiLMedUnet()
+    model = ivadomed.architecture.unet.unet.FiLMedUnet()
     inf = model(inp)
     assert (type(inf) == torch.Tensor)
 
@@ -52,7 +52,7 @@ def test_filmed_unet():
 def test_film_generator():
     a = [[[[0 for i in range(64)] for i in range(64)]]]
     inp = torch.tensor(a).float()
-    model = ivadomed.architecture.blocks.FiLMgenerator(64, 1)
+    model = ivadomed.architecture.block.film_generator.FiLMgenerator(64, 1)
     inf = model(inp)
     assert (type(inf[0]) == torch.Tensor)
     assert (type(inf[1]) == list)
