+++ testing/unit_tests/test_onnx.py
@@ -5,12 +5,11 @@ import shutil
 import logging
 
 import ivadomed.architecture.unet
-import ivadomed.architecture.unet3d
+import ivadomed.architecture.unet.unet3d
 from ivadomed import utils as imed_utils
 from ivadomed import inference as imed_inference
-from ivadomed import models as imed_models
 from pathlib import Path
-from testing.unit_tests.t_utils import create_tmp_dir,  __data_testing_dir__, download_data_testing_test_files
+from testing.unit_tests.t_utils import create_tmp_dir,  __data_testing_dir__
 from testing.common_testing_util import remove_tmp_dir
 logger = logging.getLogger(__name__)
 
@@ -27,7 +26,7 @@ LENGTH_3D = (112, 112, 112)
 
 
 def test_onnx(download_data_testing_test_files):
-    model = ivadomed.architecture.unet3d.Modified3DUNet(1, 1)
+    model = ivadomed.architecture.unet.unet3d.Modified3DUNet(1, 1)
     if not PATH_MODEL.exists():
         PATH_MODEL.mkdir()
     torch.save(model, PATH_MODEL_PT)
