+++ testing/unit_tests/test_testing.py
@@ -9,10 +9,9 @@ from ivadomed import metrics as imed_metrics
 from ivadomed import transforms as imed_transforms
 from ivadomed import utils as imed_utils
 from ivadomed import testing as imed_testing
-from ivadomed import models as imed_models
 from ivadomed.loader import utils as imed_loader_utils, loader as imed_loader
 import logging
-from testing.unit_tests.t_utils import create_tmp_dir, __data_testing_dir__, __tmp_dir__, download_data_testing_test_files, path_repo_root
+from testing.unit_tests.t_utils import create_tmp_dir, __data_testing_dir__, __tmp_dir__, path_repo_root
 from testing.common_testing_util import remove_tmp_dir
 logger = logging.getLogger(__name__)
 
@@ -92,7 +91,7 @@ def test_inference(download_data_testing_test_files, transforms_dict, test_lst, 
     })
 
     # Model
-    model = ivadomed.architecture.unet.Unet()
+    model = ivadomed.architecture.unet.unet.Unet()
 
     if cuda_available:
         model.cuda()
@@ -192,7 +191,7 @@ def test_inference_2d_microscopy(download_data_testing_test_files, transforms_di
     })
 
     # Model
-    model = ivadomed.architecture.unet.Unet(out_channel=model_params['out_channel'])
+    model = ivadomed.architecture.unet.unet.Unet(out_channel=model_params['out_channel'])
 
     if cuda_available:
         model.cuda()
